refactor: rename redis file
