+++ src/graphql/pubsub.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import { PubSubEngine } from 'graphql-subscriptions';
 import { RedisPubSub } from 'graphql-redis-subscriptions';
-import { createRedisClient } from '@/redis/client';
+import { createRedisClient } from '@/redis';
 
 // If you prefer a different PubSub implementation see:
 // https://www.apollographql.com/docs/apollo-server/data/subscriptions/#pubsub-implementations
